requirements.txt
Flask
redis
PyJWT
gunicorn
psycopg2